update to R 4.4.0

+++ pipeline-runner/R/qc-6-integrate_scdata-fastmnn.R
@@ -7,7 +7,7 @@ run_fastmnn <- function(scdata_list, config, cells_id) {
   }
 
   # calculate as many PCs for the PCA as possible, ideally 50, unless few cells
-  npcs_for_pca <- min(vapply(scdata_list, ncol, integer(1)) - 1, 50)
+  npcs_for_pca <- min(vapply(scdata_list, ncol, numeric(1)) - 1, 50)
   npcs <- config$dimensionalityReduction$numPCs
 
   # use the min of what the user wants and what can be calculated
